+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/GetUpdatesRaw.kt
@@ -1,13 +1,12 @@
 package dev.inmo.tgbotapi.extensions.api
 
 import dev.inmo.tgbotapi.bot.TelegramBot
-import dev.inmo.tgbotapi.requests.GetUpdates
 import dev.inmo.tgbotapi.requests.GetUpdatesRaw
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.update.abstracts.Update
 
 suspend fun TelegramBot.getRawUpdates(
-    offset: UpdateIdentifier? = null,
+    offset: UpdateId? = null,
     limit: Int = getUpdatesLimit.last,
     timeout: Seconds? = null,
     allowed_updates: List<String>? = ALL_UPDATES_LIST