+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/GetUpdates.kt
@@ -21,7 +21,7 @@ private val updatesListSerializer = ListSerializer(
  */
 @Serializable
 data class GetUpdates(
-    override val offset: UpdateIdentifier? = null,// set `last update id + 1` to receive next part of updates
+    override val offset: UpdateId? = null,// set `last update id + 1` to receive next part of updates
     override val limit: Int = getUpdatesLimit.last,
     override val timeout: Seconds? = null,
     override val allowed_updates: List<String>? = ALL_UPDATES_LIST