+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/GetUpdatesRequest.kt
@@ -1,14 +1,11 @@
 package dev.inmo.tgbotapi.requests
 
-import dev.inmo.tgbotapi.requests.abstracts.Request
 import dev.inmo.tgbotapi.requests.abstracts.SimpleRequest
-import dev.inmo.tgbotapi.types.ALL_UPDATES_LIST
 import dev.inmo.tgbotapi.types.Seconds
-import dev.inmo.tgbotapi.types.UpdateIdentifier
-import dev.inmo.tgbotapi.types.getUpdatesLimit
+import dev.inmo.tgbotapi.types.UpdateId
 
 interface GetUpdatesRequest<T : Any> : SimpleRequest<T> {
-    val offset: UpdateIdentifier?
+    val offset: UpdateId?
     val limit: Int
     val timeout: Seconds?
     val allowed_updates: List<String>?