+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/update/CommonChatMemberUpdatedUpdate.kt
@@ -1,10 +1,10 @@
 package dev.inmo.tgbotapi.types.update
 
 import dev.inmo.tgbotapi.types.chat.member.ChatMemberUpdated
-import dev.inmo.tgbotapi.types.UpdateIdentifier
+import dev.inmo.tgbotapi.types.UpdateId
 import dev.inmo.tgbotapi.types.update.abstracts.ChatMemberUpdatedUpdate
 
 data class CommonChatMemberUpdatedUpdate(
-    override val updateId: UpdateIdentifier,
+    override val updateId: UpdateId,
     override val data: ChatMemberUpdated
 ) : ChatMemberUpdatedUpdate