+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   * `WebAppQueryId` now is `value class`
   * `PreCheckoutQueryId` now is `value class`
   * `FileUniqueId` has been renamed to `TgFileUniqueId` and now is `value class`
+  * `UpdateIdentifier` has been renamed to `UpdateId` and now is `value class`
 
 ## 10.1.2